[EI-244] [api] refactor: 파일명 변경, 폴더 구조 변경
