+++ api/src/numerical-guidance/api/indicator/dto/search-twelve-symbol.dto.ts
