--- api/src/numerical-guidance/application/port/persistence/indicator/search-indicator.port.ts
@@ -1,5 +1,0 @@
-import { SearchedIndicatorsDto } from '../../../query/indicator/get-indicator-search/dto/searched-indicators.dto';
-
-export interface SearchIndicatorPort {
-  searchIndicator(symbol: string): Promise<SearchedIndicatorsDto>;
-}
