+++ api/src/numerical-guidance/application/port/persistence/indicator/search-twelve-indicator.port.ts
@@ -1,0 +1,5 @@
+import { SearchedIndicatorsDto } from '../../../query/indicator/search-twelve-indicator/dto/searched-indicators.dto';
+
+export interface SearchTwelveIndicatorPort {
+  searchIndicator(symbol: string): Promise<SearchedIndicatorsDto>;
+}
