+++ api/src/numerical-guidance/application/query/custom-forecast-indicator/get-custom-forecast-indicator-values/get-custom-forecast-indicator-values.query.handler.ts
@@ -65,7 +65,6 @@ export class GetCustomForecastIndicatorValuesQueryHandler implements IQueryHandl
       customForecastIndicatorValues: customFroecastIndicatorValues.indicatorValues,
       targetIndicatorValues: targetIndicatorValues,
     };
-    console.log(customForecastIndicatorValuesResponse);
     return customForecastIndicatorValuesResponse;
   }
