--- api/src/numerical-guidance/application/query/indicator/get-search-indicator-by-symbol/search-indicator-by-symbol.query.ts
@@ -1,5 +1,0 @@
-import { IQuery } from '@nestjs/cqrs';
-
-export class SearchIndicatorBySymbolQuery implements IQuery {
-  constructor(readonly symbol: string) {}
-}
