+++ api/src/numerical-guidance/application/query/indicator/search-twelve-indicator/dto/searched-indicators.dto.ts
