+++ api/src/numerical-guidance/application/query/indicator/search-twelve-indicator/search-twelve-indicator.query.handler.ts
@@ -1,0 +1,19 @@
+import { IQueryHandler, QueryHandler } from '@nestjs/cqrs';
+import { Inject, Injectable } from '@nestjs/common';
+import { SearchTwelveIndicatorQuery } from './search-twelve-indicator.query';
+import { SearchTwelveIndicatorPort } from '../../../port/persistence/indicator/search-twelve-indicator.port';
+import { SearchedIndicatorsDto } from './dto/searched-indicators.dto';
+
+@Injectable()
+@QueryHandler(SearchTwelveIndicatorQuery)
+export class SearchTwelveIndicatorQueryHandler implements IQueryHandler {
+  constructor(
+    @Inject('SearchTwelveIndicatorPort')
+    private readonly searchTwelveIndicatorPort: SearchTwelveIndicatorPort,
+  ) {}
+
+  async execute(query: SearchTwelveIndicatorQuery): Promise<SearchedIndicatorsDto> {
+    const { symbol } = query;
+    return this.searchTwelveIndicatorPort.searchIndicator(symbol);
+  }
+}
